Fix channel reciver thread safe problem
Taking a job off the channel queue involves mutating receiver, so the threads
need a safe way to share receiver and be allowed to modify it. If the
modifications weren’t threadsafe, we might get race conditions such as two
threads executing the same job if they both take the same job off the queue at the same time.

ref: https://doc.rust-lang.org/1.24.0/book/second-edition/ch20-05-sending-requests-via-channels.html